Add extension dependencies support to event based extensions
